Make README use keep-alive-interval MQTT connect option
This connection option is more probable to be used in real world
configurations than the max-inflight option shown before. Use
keep-alive-interval for the example configurations, so non-expert MQTT
users copyping the configuration from the README file know about it
and can tune it for their setup. Or at least have a reasonable value
for always connected MQTT clients, so firewalls and other middle-boxes
don't tear the TCP connection down due to prolonged inactivity.