Add client-id parameter to broker configuration
In many cases, specifying the client-id for the MQTT connection is not
important. Usually a random client-id is generated and used. But in
certain cases being able to specify which exact client-id is
needed. E.g., when the client-id is part of the topic name and used
for access control.

[Re: #1]