Create new certificates for local integration tests
The old ones for the client and the server had expired, so tests were
failing when done locally with "self-signed" certificates.

Also commit all the certification chain, from the top Root CA down to
the client and server certificates. We don't expect to need to renew
them ever (they last for 20 years) but one never knows.

Given that they are not used anywhere except in integration tests run
locally in the developer machines, and all the services that use them
are absolutely ephemeral, this is not a security issue.