Fix support for :opts key in AMQP broker configuration
README.md documents that there is an optional :opts key in the broker
configuration that can be used to pass additional AMQP connection
options to the underlying library.

But the whole :opts key (and associated value) is passed in as another
AMQP connection parameter. Which means the underlying AMQP library
completely ignores it (as it doesn't use such :opts key at all).

[Fixes: #6]