Fix tests for AMQP event listeners
The shutdown event listener didn't record the root cause in some cases
(it was nil for some types of network errors, e.g., when the broker
was shutdown but the AMQP client was still up and running). So change
the code to get something at least, even if in some cases we could get
a more detailed root cause (in some cases the exception stack has a
better root cause diagnostic in the next entry of the stack trace, but
that isn't guaranteed to always be that way).

The recovery listener threw an exception while trying to build a
keyword for the log entry (because it was trying to do things that
weren't right, mainly trying to use non-string values as strings). And
as a consequence, the whole connection recovery dies and the AMQP
client library never recovers the connection. Because this use case
was never exercised from the unit tests, we hadn't noticed it.